refactor: consolidate UI layer into single ui/ directory
