+++ nx/blocks/media-library/ui/views/mediainfo/mediainfo.css
