+++ nx/blocks/media-library/ui/views/onboard/onboard.js
@@ -1,12 +1,12 @@
 import { html, LitElement, nothing } from 'da-lit';
-import getStyle from '../../../../utils/styles.js';
-import { parseOrgRepoFromUrl } from '../../core/urls.js';
-import { normalizeSitePath } from '../../core/paths.js';
-import loadSvgIcons from '../../../../utils/svg.js';
-import { Storage } from '../../core/constants.js';
-import { showNotification } from '../../core/state.js';
-import { t } from '../../core/messages.js';
-import { ErrorCodes, logMediaLibraryError } from '../../core/errors.js';
+import getStyle from '../../../../../utils/styles.js';
+import { parseOrgRepoFromUrl } from '../../../core/urls.js';
+import { normalizeSitePath } from '../../../core/paths.js';
+import loadSvgIcons from '../../../../../utils/svg.js';
+import { Storage } from '../../../core/constants.js';
+import { showNotification } from '../../../core/state.js';
+import { t } from '../../../core/messages.js';
+import { ErrorCodes, logMediaLibraryError } from '../../../core/errors.js';
 
 const EL_NAME = 'nx-media-onboard';
 const styles = await getStyle(import.meta.url);
