+++ nx/blocks/media-library/ui/views/sidebar/sidebar.css
