+++ nx/blocks/media-library/ui/views/sidebar/sidebar.js
@@ -1,8 +1,8 @@
 import { html, LitElement } from 'da-lit';
-import getStyle from '../../../../utils/styles.js';
-import getSvg from '../../../../public/utils/svg.js';
-import { getAppState, onStateChange } from '../../core/state.js';
-import { t } from '../../core/messages.js';
+import getStyle from '../../../../../utils/styles.js';
+import getSvg from '../../../../../public/utils/svg.js';
+import { getAppState, onStateChange } from '../../../core/state.js';
+import { t } from '../../../core/messages.js';
 
 const styles = await getStyle(import.meta.url);
 const nx = `${new URL(import.meta.url).origin}/nx`;
